requirements.txt
platformdirs==4.2.2
PyYAML==6.0.1
PyYAML==6.0.1
setuptools==70.2.0
textual==0.71.0
